arch: split event_handler into multiple files
